@@ -4,8 +4,8 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "litprinter"
-version = "0.2.2"
-description = "The most sophisticated debug printing library for Python with rich formatting, syntax highlighting, and beautiful tracebacks"
+dynamic = ["version"]
+description = "The most sophisticated debug printing library for Python - IceCream + Rich fusion with syntax highlighting and beautiful tracebacks"
 readme = "README.md"
 authors = [
     {name = "OEvortex", email = "helpingai5@gmail.com"}
@@ -16,22 +16,23 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.6",
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Software Development :: Debuggers",
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
-keywords = ["debug", "print", "logging", "traceback", "formatting"]
+keywords = ["debug", "print", "icecream", "rich", "logging", "traceback", "formatting", "syntax-highlighting"]
 dependencies = [
     "pygments",
     "colorama",
     "executing",
+    "asttokens",
 ]
-requires-python = ">=3.6"
+requires-python = ">=3.8"
 
 [project.urls]
 "Homepage" = "https://github.com/OEvortex/litprinter"
@@ -54,16 +55,19 @@ package-dir = {"" = "src"}
 [tool.setuptools.packages.find]
 where = ["src"]
 
+[tool.setuptools.dynamic]
+version = {attr = "litprinter.__version__"}
+
 [tool.black]
 line-length = 88
-target-version = ["py36", "py37", "py38", "py39", "py310", "py311"]
+target-version = ["py38", "py39", "py310", "py311", "py312"]
 
 [tool.isort]
 profile = "black"
 line_length = 88
 
 [tool.mypy]
-python_version = "3.6"
+python_version = "3.8"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = false
